Diagnostics dump carries the memory split and the full profiler report
Every dump now reports anonymous, file-backed and shared RSS plus the cgroup
figures next to the total, so mapped database pages and cached files can be
told from heap growth without installing anything. While the Profiler is
installed the dump embeds its full report, with a day of recorder history
thinned to one sample per five minutes; sections can declare their own
timeout for that. Periodic CPU profile windows become opt-in: Python runs
about 4.5 times slower while yappi captures one, enough to disturb playback.
