+++ music_assistant/translations/en.json
@@ -2059,7 +2059,7 @@
   "provider.profiler.config_entries.action_run_cpu_profile.label": "Capture an on-demand CPU profile window, for example while you are reproducing an issue. The result is included in the next report.",
   "provider.profiler.config_entries.cpu_profile_duration.description": "How long each CPU profile window records before its results are processed.",
   "provider.profiler.config_entries.cpu_profile_duration.label": "CPU profile window duration (seconds)",
-  "provider.profiler.config_entries.cpu_profile_enabled.description": "Automatically capture a CPU profile window at a regular interval. The server runs slightly slower while a window is being captured.",
+  "provider.profiler.config_entries.cpu_profile_enabled.description": "Automatically capture a CPU profile window at a regular interval. Python code runs several times slower while a window is being captured, which can disturb playback on a busy server - prefer the on-demand button unless you are hunting a CPU problem.",
   "provider.profiler.config_entries.cpu_profile_enabled.label": "Periodic CPU profiling",
   "provider.profiler.config_entries.cpu_profile_interval.description": "How often a new CPU profile window is captured automatically.",
   "provider.profiler.config_entries.cpu_profile_interval.label": "CPU profile interval (minutes)",
